Initial commit - package hiera-eyaml for Puppet 5 on Ubuntu trusty
This will allow the building of hiera-eyaml for Puppet 5 on Ubuntu trusty,
which has its own bundled ruby in /opt/puppetlabs/puppet. These packages will
install the gems there, which is a slight violation of the FHS and general
Ubuntu packaging principles, but no more so than the puppet packages
themselves.